+++ src/classes/utils/xmlmetadata/XMLSourcePrompt.ts
@@ -1,6 +1,12 @@
 import { AxiosRequestConfig } from "axios";
 import { HTTPHandler } from "ui5plugin-parser/dist/classes/http/HTTPHandler";
 import * as vscode from "vscode";
+
+export interface TSODataInterfacesFetchingData {
+	username?: string;
+	password?: string;
+	url?: string;
+}
 export class XMLSourcePrompt {
 	async getXMLMetadataText() {
 		let XMLMetadata = "";
@@ -8,23 +14,37 @@ export class XMLSourcePrompt {
 		if (activeDocument && activeDocument.fileName.endsWith("metadata.xml")) {
 			XMLMetadata = activeDocument.getText();
 		} else {
-			const uri = await vscode.window.showInputBox({
-				prompt: "Please define url to metadata"
-			});
-			const username = await vscode.window.showInputBox({
-				prompt: "Please enter username"
-			});
-			const password = await vscode.window.showInputBox({
-				prompt: "Please enter password",
-				password: true
-			});
+			const config = vscode.workspace
+				.getConfiguration("ui5.plugin")
+				.get<TSODataInterfacesFetchingData | undefined>("TSODataInterfacesFetchingData");
 
-			if (uri) {
+			let url: string | undefined;
+			let username: string | undefined;
+			let password: string | undefined;
+
+			if (config?.url) {
+				url = config.url;
+				username = config.username;
+				password = config.password;
+			} else {
+				url = await vscode.window.showInputBox({
+					prompt: "Please define url to metadata"
+				});
+				username = await vscode.window.showInputBox({
+					prompt: "Please enter username"
+				});
+				password = await vscode.window.showInputBox({
+					prompt: "Please enter password",
+					password: true
+				});
+			}
+
+			if (url) {
 				const config: AxiosRequestConfig = {};
 				if (username && password) {
 					config.auth = { username, password };
 				}
-				XMLMetadata = await HTTPHandler.get(uri, config);
+				XMLMetadata = await HTTPHandler.get(url, config);
 			}
 		}